+++ game.js
@@ -69,6 +69,7 @@ var Game = (function() {
         self.minerUI.update(delta);
         self.colonyUI.update(delta);
         self.laboratoryUI.update(delta);
+        self.planetUI.update(delta);
         self.goldenEventUI.update(delta);
 
         self.updateAutoSave(delta);
@@ -147,6 +148,7 @@ var Game = (function() {
         this.settings.save(data);
         this.miners.save(data);
         this.laboratory.save(data);
+        this.planets.save(data);
         this.goldenEvents.save(data);
         this.interstellar.save(data);
         this.stargaze.save(data);
@@ -171,6 +173,7 @@ var Game = (function() {
             this.buildings.load(data);
             this.miners.load(data);
             this.laboratory.load(data);
+            this.planets.load(data);
             this.goldenEvents.load(data);
             this.stargaze.load(data);
             this.tech.load(data);
@@ -254,6 +257,7 @@ var Game = (function() {
         self.stargaze.initialise();
         self.miners.initialise();
         self.laboratory.initialise();
+        self.planets.initialise();
         self.goldenEvents.initialise();
 
         // Now load
@@ -263,6 +267,7 @@ var Game = (function() {
         self.minerUI.initialise();
         self.colonyUI.initialise();
         self.laboratoryUI.initialise();
+        self.planetUI.initialise();
         self.goldenEventUI.initialise();
 
         for(var i = 0; i < self.uiComponents.length; i++) {